🔨 Refactor(#193): StoreMyCharacterSection 및 SortBar 컴포넌트 경로 수정 및 불필요한 파일 삭제
