+++ src/features/store/ui/StoreMyCharacterSection.tsx
@@ -1,11 +1,11 @@
 import CartList from '@/features/store/ui/CartList';
 import MyCharacter from '@/features/user/ui/MyCharacter';
 import { MyCharacterSection, StoreButton } from '@/pages/store/styles';
-import useModal from './../../../hooks/useModal';
+import useModal from '../../../hooks/useModal';
 import { userCosmeticItemsQuery } from '@/features/user/queries';
 import { useCosmeticItemStore } from '@/features/store/store';
 import ProfileImage from '@/features/user/ui/ProfileImage';
-import useUserStore from './../../../store/useUserStore';
+import useUserStore from '../../../store/useUserStore';
 import { isLoggedIn } from '@/features/user/service/authUtils';
 import PreviewMyCharacter from '@/features/store/ui/PreviewMyCharacter';
 import PreViewProfileImage from '@/features/store/ui/PreviewMyProfileImage';
